requirements.txt
requests
python-telegram-bot>=11,<13
grpcio
grpcio-tools
